seconed commit

+++ src/main/java/com/flexon/calc/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator
 
 	}
 	
-	public static double subtract(double a, double b) {
+	public static int subtract(int a, int b) {
 		
 		return a-b;
 	}
